Refactored backend + fixed minor frontend bug with errors
